feat(glm5next-w2/310p): GLM-5.3-Flash-W2 TP4 bring-up — fits, custom Cube kernel (~17x decode), partial correctness
Loads + fits GLM-5.3-Flash-W2 (Glm5NextW2ForCausalLM) at TP4 on 4x Ascend 310P
(~25.5GB/chip, no RAM upgrade) and generates. NOT yet coherent — see caveat.

Fit / functionality:
- Expert parallelism for the 288 2-bit routed experts (72/rank + all-reduce);
  they were replicated -> 41GB OOM. (models/glm5next_w2/moe.py, model.py)
- 310P KV-cache init for the hybrid stack: MultipleOf coercion, dispatch by
  spec type + UniformTypeKVCacheSpecs unwrap, DSA indexer buffer grow-on-demand,
  MLA enable path. (_310p/model_runner_310p.py, attention/indexer_kpool.py,
  core/kv_cache_interface.py, _310p/attention/mla_v1_310.py, envs.py, platform.py,
  patch/platform/patch_mamba_config_310.py, ops/rotary_embedding.py)

Custom AscendC Cube kernel (the real speed win):
- npu_w2_blocked_dequant_matmul_310 (csrc/gmm/w2_blocked_dequant_matmul_v310/):
  arch20 catlass block_mmad, MIX AIC/AIV, per-[32,32] 2-bit dequant fused on-chip
  from PACKED codes. Stock npu_quant_grouped_matmul_dequant is per-channel-scale
  only -> can't express W2's along-K block scale. Parity rel<=7e-4; ~17x faster
  decode than eager (16 tok 16.9s vs 295s). Integrated in w2_dynamic._apply_device
  (eager fp32 fallback via VLLM_ASCEND_W2_DISABLE_CUBE=1).
  (csrc/build_aclnn.sh, torch_binding{,_meta}.cpp, _310p/quantization/methods/w2_dynamic.py)

Correctness (partial — output is real English words, not yet coherent):
- Dead FFN: dense-MLP + shared-expert F16 weights were wrongly fp8-quantized
  (no weight_scale_inv in ckpt) -> zeroed; mark them unquantized. (glm5next_w2/model.py)
- Residual fp16 overflow (65504 sat): carry mHC residual in fp32. (glm5next/model.py,
  patch/worker/patch_triton.py; opensensor vllm patch in artifacts/opensensor-vllm-patches/)
- KDA under-contribution: o_norm eps swamped the tiny recurrence -> lower eps
  (workaround; proper fix is fp32 recurrence). (glm5next_w2/kda.py)
- Remaining tail (documented in artifacts/glm-w2-device-bringup-log.md): near-degenerate
  /nondeterministic logits; root theme = GLM is bf16-designed, 310P is fp16-only.

+++ vllm_ascend/patch/worker/patch_triton.py
@@ -242,9 +242,14 @@ try:
     from vllm.model_executor.layers import mhc as _mhc_mod
 
     def _mhc_rms_norm(x, weight, eps):
+        # `x` (the mHC layer_input) may arrive in fp32 because the residual
+        # accumulator rides fp32 on the 310P. RMSNorm normalizes away the
+        # magnitude, so downcast the result to the norm weight's compute dtype
+        # (fp16) for the attn/MLP submodules; the fp32 accumulator is carried
+        # separately by mhc_post and is never touched here.
         xf = x.float()
         var = xf.square().mean(dim=-1, keepdim=True)
-        return (xf * _t.rsqrt(var + eps) * weight.float()).to(x.dtype)
+        return (xf * _t.rsqrt(var + eps) * weight.float()).to(weight.dtype)
 
     def _mhc_pre_npu(
         self,
